feat: add CSRF protection to all forms

+++ app/Controllers/CategoryController.php
@@ -22,17 +22,29 @@ class CategoryController extends Controller {
 
     public function store() {
         $this->adminOnly();
+        if (!$this->verifyCsrfToken()) {
+            header('Location: /');
+            exit();
+        }
+
         $name = $_POST['name'];
         $slug = strtolower(str_replace(' ', '-', $name));
+
         $this->category->create($name, $slug);
         $this->log('Category created');
+
         header('Location: /categories');
         exit();
     }
     
     public function destroy($id) {
         $this->adminOnly();
+        if (!$this->verifyCsrfToken()) {
+            header('Location: /');
+            exit();
+        }
         $this->category->delete($id);
+        
         header('Location: /categories');
         exit();
     }
